DQN/requirements.txt
pygame 
numpy
neat-python